Major fixes to handling of UTC time.
This fixes a number of bugs in calculating and using UTC time when the machine, bedework and user timezone settings differ.

Created a new FlushMap class (extracted from core directory classes) and used it to flush timezone information on a regular basis.

Fix caching of calculated UTC for date values.

Stop passing timezone as a parameter (back to ID only)